Refactor: extract config sync hook, unify settings types, add generic RPC client
Three code simplifications identified in architecture review:
1. Extract ~200 lines of config plumbing from page.tsx into useConfigSync hook
2. Make WorkspacePreferences = Required<CushionSettings> (eliminates triple field mapping)
3. Add typed call() method to CoordinatorClient (convenience methods become thin wrappers)
